+++ KFXAdditions/Classes/UIKit/UIView/UIView+KFXAdditions.m
@@ -131,6 +131,21 @@
 	
 }
 
+-(void)kfx_addRoundedCorners:(UIRectCorner)corners radius:(CGFloat)radius{
+    
+    CGRect rect = self.bounds;
+    
+    // Create the path
+    UIBezierPath *maskPath = [UIBezierPath bezierPathWithRoundedRect:rect byRoundingCorners:corners cornerRadii:CGSizeMake(radius, radius)];
+    
+    // Create the shape layer and set its path
+    CAShapeLayer *maskLayer = [CAShapeLayer layer];
+    maskLayer.frame = rect;
+    maskLayer.path = maskPath.CGPath;
+    
+    // Set the newly created shape layer as the mask for the view's layer
+    self.layer.mask = maskLayer;
+}
 
 //--------------------------------------------------------
 #pragma mark Animation